Account creation #3: Links (#1504)
### Motivation
From [issue](https://linear.app/careerfairy/issue/CF-1195/add-options-to-share-different-links-on-account-creation): In order to be able to release the new talent profile we need to update the existing account creation experience to incorporate the data changes.

This ticket ensures giving the option for users to add different links to their profile upon account creation. 

<!-- A little context about this change. -->

### Changes

-  Updated `SocialInformation.tsx`, added custom link form.

**There is already an hotfix for this file, meaning conflicts must be resolved before merging**

<!-- You can remove anything that doesn't make sense for this change. -->

### Testing
1. Create a new account using https://careerfairy-ssr-webapp-pr-1504.vercel.app/signup?talentProfileV1=true 
2. In the social section you can now add additional links and when entering a LinkedIn link its displayed in the same style as additional links.
### Demo

<!-- Consider including a short demo video. -->

<div class='graphite__hidden'>
          <div>🎥 Video uploaded on Graphite:</div>
            <a href="https://app.graphite.dev/media/video/S0WylOadBGaro8NXGBwS/fb92e5f9-918e-468f-8883-2ce74a41681a.mov">
              <img src="https://app.graphite.dev/api/v1/graphite/video/thumbnail/S0WylOadBGaro8NXGBwS/fb92e5f9-918e-468f-8883-2ce74a41681a.mov">
            </a>
          </div>
<video src="https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/S0WylOadBGaro8NXGBwS/fb92e5f9-918e-468f-8883-2ce74a41681a.mov">Gravação do ecrã 2025-01-20, às 10.36.34.mov</video>

### Browser Testing

-  [x] Chrome
-  [ ] Firefox
-  [ ] Safari

### Different Resolutions Testing

-  [x] Large Screens
-  [ ] Medium Screens
-  [x] Small Screens

+++ apps/web/layouts/UserLayout/TalentProfile/Details/Profile/ProfileLinks.tsx
@@ -5,12 +5,13 @@ import { useAuth } from "HOCs/AuthProvider"
 import useIsMobile from "components/custom-hook/useIsMobile"
 import useSnackbarNotifications from "components/custom-hook/useSnackbarNotifications"
 import { useUserLinks } from "components/custom-hook/user/useUserLinks"
+import { LINKEDIN_URL_REGEX } from "components/util/constants"
 import CircularLogo from "components/views/common/logos/CircularLogo"
 import { userRepo } from "data/RepositoryInstances"
 import Link from "next/link"
 import normalizeUrl from "normalize-url"
 import { OptionsObject } from "notistack"
-import { Fragment, useCallback, useState } from "react"
+import { Fragment, useCallback, useMemo, useState } from "react"
 import { ExternalLink, Link as FeatherLink } from "react-feather"
 import { useFormContext } from "react-hook-form"
 import { useDispatch, useSelector } from "react-redux"
@@ -144,10 +145,35 @@ const FormDialogWrapper = () => {
             authId: userData.authId,
          }
 
-         if (!data?.id) {
-            await userRepo.createUserLink(userData.id, newLink)
+         // User has no linkedin url and the new url is valid (a linkedin url)
+         if (!userData.linkedinUrl && LINKEDIN_URL_REGEX.test(data.url)) {
+            await userRepo.updateAdditionalInformation(userData.id, {
+               linkedinUrl: data.url,
+            })
          } else {
-            await userRepo.updateUserLink(userData.id, newLink)
+            if (data?.id === "linkedin") {
+               // User is updating their linkedin url, leaving it as a valid linkedin url or
+               // changing it to an invalid linkedin url but still keeping it in the user links.
+
+               // Either way the linkedinUrl must be updated, empty if the url is not valid and the url if it is valid.
+               const isValidLinkedinUrl = LINKEDIN_URL_REGEX.test(data.url)
+
+               await userRepo.updateAdditionalInformation(userData.id, {
+                  linkedinUrl: isValidLinkedinUrl ? data.url : "",
+               })
+
+               if (!isValidLinkedinUrl) {
+                  // Since the linkedin url is not valid anymore, we create a new user link
+                  await userRepo.createUserLink(userData.id, newLink)
+               }
+            } else {
+               // User is creating/updating a non-linkedin link
+               if (!data?.id) {
+                  await userRepo.createUserLink(userData.id, newLink)
+               } else {
+                  await userRepo.updateUserLink(userData.id, newLink)
+               }
+            }
          }
 
          handleCloseLinkDialog()
@@ -189,12 +215,28 @@ const FormDialogWrapper = () => {
 }
 
 const LinksList = () => {
-   const { data: userLinks } = useUserLinks()
+   const { userData } = useAuth()
+
+   const { data: userStoredLinks } = useUserLinks()
    const dispatch = useDispatch()
 
    const handleAdd = useCallback(() => {
       dispatch(openCreateDialog({ type: TalentProfileItemTypes.Link }))
    }, [dispatch])
+
+   const userLinks = useMemo(() => {
+      return userData.linkedinUrl
+         ? [
+              {
+                 title: "LinkedIn",
+                 url: userData.linkedinUrl,
+                 authId: userData.authId,
+                 id: "linkedin",
+              },
+              ...userStoredLinks,
+           ]
+         : userStoredLinks
+   }, [userStoredLinks, userData.linkedinUrl, userData.authId])
 
    if (!userLinks?.length)
       return (
@@ -248,7 +290,13 @@ const LinkCard = ({ link }: LinkCardProps) => {
    const handleDelete = useCallback(async () => {
       setIsDeleting(true)
 
-      await userRepo.deleteLink(userData.id, link.id)
+      if (link.id === "linkedin") {
+         await userRepo.updateAdditionalInformation(userData.id, {
+            linkedinUrl: null,
+         })
+      } else {
+         await userRepo.deleteLink(userData.id, link.id)
+      }
 
       setIsDeleting(false)
       setIsConfirmDeleteDialogOpen(false)
@@ -327,4 +375,4 @@ const LinkCard = ({ link }: LinkCardProps) => {
    )
 }
 
-const isLinkedInUrl = (url: string) => url.includes("linkedin.")
+export const isLinkedInUrl = (url: string) => url.includes("linkedin.")
